+++ src/main.ts
@@ -1,5 +1,6 @@
-import { App, Editor, MarkdownFileInfo, MarkdownView, Modal, Notice, Plugin } from 'obsidian';
+import { Plugin } from 'obsidian';
 import { isVersionNewerThanOther } from './changelog';
+import { installNameGuard } from './name-guard';
 import { ReleaseNotesModal } from './release-notes-modal';
 import { DEFAULT_SETTINGS, NameGuardPluginSettings, NameGuardSettingTab } from './settings';
 
@@ -9,52 +10,9 @@ export default class NameGuardPlugin extends Plugin {
 	async onload() {
 		await this.loadSettings();
 
-		// This creates an icon in the left ribbon.
-		this.addRibbonIcon('dice', this.manifest.name, (evt: MouseEvent) => {
-			// Called when the user clicks the icon.
-			new Notice(`This is a ${this.manifest.name} notice!`);
-		});
-
-		// This adds a status bar item to the bottom of the app. Does not work on mobile apps.
-		const statusBarItemEl = this.addStatusBarItem();
-		statusBarItemEl.setText('Status bar text');
-
-		// This adds a simple command that can be triggered anywhere
-		this.addCommand({
-			id: 'open-modal-simple',
-			name: 'Open modal (simple)',
-			callback: () => {
-				new NameGuardModal(this.app).open();
-			},
-		});
-		// This adds an editor command that can perform some operation on the current editor instance
-		this.addCommand({
-			id: 'replace-selected',
-			name: 'Replace selected content',
-			editorCallback: (editor: Editor, view: MarkdownView | MarkdownFileInfo) => {
-				editor.replaceSelection('NameGuard editor command');
-			},
-		});
-		// This adds a complex command that can check whether the current state of the app allows execution of the command
-		this.addCommand({
-			id: 'open-modal-complex',
-			name: 'Open modal (complex)',
-			checkCallback: (checking: boolean) => {
-				// Conditions to check
-				const markdownView = this.app.workspace.getActiveViewOfType(MarkdownView);
-				if (markdownView) {
-					// If checking is true, we're simply "checking" if the command can be run.
-					// If checking is false, then we want to actually perform the operation.
-					if (!checking) {
-						new NameGuardModal(this.app).open();
-					}
-
-					// This command will only show up in Command Palette when the check function returns true
-					return true;
-				}
-				return false;
-			},
-		});
+		// Wrap the vault mutation methods so duplicate names are blocked before
+		// a file is ever created. Wrappers are restored on unload.
+		installNameGuard(this);
 
 		this.addCommand({
 			id: 'show-release-notes',
@@ -64,14 +22,7 @@ export default class NameGuardPlugin extends Plugin {
 			},
 		});
 
-		// This adds a settings tab so the user can configure various aspects of the plugin
 		this.addSettingTab(new NameGuardSettingTab(this.app, this));
-
-		// If the plugin hooks up any global DOM events (on parts of the app that doesn't belong to this plugin)
-		// Using this function will automatically remove the event listener when this plugin is disabled.
-		// this.registerDomEvent(activeDocument, 'click', (evt: MouseEvent) => {
-		// 	new Notice('Click');
-		// });
 
 		this.app.workspace.onLayoutReady(() => this.maybeShowReleaseNotes());
 	}
@@ -101,21 +52,5 @@ export default class NameGuardPlugin extends Plugin {
 
 	async saveSettings() {
 		await this.saveData(this.settings);
-	}
-}
-
-class NameGuardModal extends Modal {
-	constructor(app: App) {
-		super(app);
-	}
-
-	onOpen() {
-		let { contentEl } = this;
-		contentEl.setText('Woah!');
-	}
-
-	onClose() {
-		const { contentEl } = this;
-		contentEl.empty();
 	}
 }